backend/requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
flask-ngrok==0.0.25
idna==3.4
imagetools==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
ngrok==0.12.0
piexif==1.1.3
Pillow==10.1.0
requests==2.31.0
urllib3==2.0.7
Werkzeug==3.0.0
